Original Version - Draft 3
Quantifies the ratio of primary+secondary staining to secondary only
staining as a metric of primary specificity.   The data  is expected to
be in a format of alternating primary and secondary only images/stacks,
and that each image set contains at least an autofluroescence channel
and an antibody channel (additional channels are acceptable).